Удален react context, установлен redux, react-redux. Сделаны контейнерные компоненты
