+++ package.json
@@ -13,9 +13,11 @@
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "react-scrollable-feed": "^1.1.2",
+    "redux": "^4.0.5",
     "sass-loader": "^10.0.1"
   },
   "scripts": {
