+++ src/components/Login.js
@@ -1,52 +1,12 @@
-import React, {useState} from 'react'
-import {auth, provider} from '../firebase/firebase'
-import {useStateValue} from './../StateProvider'
-import {actionTypes} from '../reducer'
+import React from 'react'
 import {Link} from 'react-router-dom'
-import db from './../firebase/firebase'
 
-const Login = () => {
-  const [_, dispatch] = useStateValue()
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-
-  const signInWithGoogle = (e) => {
-    e.preventDefault()
-    auth
-      .signInWithPopup(provider)
-      .then((result) => {
-        dispatch({
-          type: actionTypes.SET_USER,
-          user: result.user,
-        })
-        db.collection('users').doc(result.user.uid).set(
-          {
-            displayName: result.user.displayName,
-            photoURL: result.user.photoURL,
-            uid: result.user.uid,
-            email: result.user.email,
-          },
-          {merge: true}
-        )
-      })
-      .catch((e) => alert(e.message))
-  }
-
-  const signInWithPassword = (e) => {
-    e.preventDefault()
-    auth
-      .signInWithEmailAndPassword(email, password)
-      .then((result) => {
-        db.collection('users').doc(result.user.uid).set({
-          displayName: result.user.displayName,
-          photoURL: result.user.photoURL,
-          uid: result.user.uid,
-          email: result.user.email,
-        })
-      })
-      .catch((err) => alert(err))
-  }
-
+const Login = ({
+  signInWithPassword,
+  signInWithGoogle,
+  setEmail,
+  setPassword,
+}) => {
   return (
     <div className="login auth__wrapper">
       <main>
