+++ src/types.js
@@ -1,0 +1,1 @@
+export const SET_USER_DATA = 'SET_USER_DATA'
